Separation(code)
Images, for representing the frame data, a numerical matrix whose values can be colors, color
map entries, etc. A shot object is a discrete sequence of images with a set of temporal attributes such as frame rate and duration
and represents a video segment. A shot sequence object groups several shots using some semantic criteria.
Views, are used to visualize and browse shots and shot sequences.